Avoid accidents and adding empty data

+++ lib/tasks/reimport.rake
@@ -1,6 +1,8 @@
 namespace :reimport do
   desc 'Reimport TPI data - USE WITH CAUTION'
   task tpi: :environment do
+    next if Rails.env.production?
+
     Company.delete_all
     CP::Assessment.delete_all
     MQ::Assessment.delete_all
@@ -10,6 +12,8 @@ namespace :reimport do
   end
 
   task tpi_sector_clusters: :environment do
+    next if Rails.env.production?
+
     Seed::TPIData.import_sector_clusters
 
     puts 'TPI Sector Clusters reimported'
@@ -17,6 +21,8 @@ namespace :reimport do
 
   desc 'Reimport CCLOW data - USE WITH CAUTION'
   task cclow: :environment do
+    next if Rails.env.production?
+
     Legislation.delete_all
     Keyword.delete_all
     Framework.delete_all
@@ -33,13 +39,17 @@ namespace :reimport do
 
   desc 'Reimport CCLOW files data - USE WITH CAUTION'
   task cclow_sources: :environment do
+    next if Rails.env.production?
+
     Document.delete_all
 
     Seed::CCLOWData.call_sources_import
   end
 
   desc 'Reimport CCLOW litigation files data only - USE WITH CAUTION'
-  task cclow_sources: :environment do
+  task cclow_litigation_sources: :environment do
+    next if Rails.env.production?
+
     # let's start by appending
     # Document.delete_all
 
@@ -48,6 +58,8 @@ namespace :reimport do
 
   desc 'Reimport CCLOW LitigationSides'
   task cclow_litigation_sides: :environment do
+    next if Rails.env.production?
+
     LitigationSide.delete_all
 
     Seed::CCLOWData.import_litigation_sides
